requirements.txt
asgiref==3.3.4
boto3==1.17.78
Django==3.2.3
django-filter==2.4.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1